LP-67: fixing the required field priceCourse on course.html to accept decimal number

+++ src/main/java/com/java/study/group/librarysystem/dto/CourseDto.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -16,7 +17,7 @@ public class CourseDto {
     private int limitOfCostumers;
     private LocalDateTime startDateAndTime;
     private String ageGroup;
-    private Long priceOfClass;
+    private BigDecimal priceOfClass;
     private String instructorName;
 
     public Course toCourse(){
